requirements-ci.txt
# --- 核心依赖 ---
# 注意：这里不写 torch 和 torchvision，因为我们在 YAML 里单独用命令装 CPU 版
numpy<2.0.0
ipython
opencv-python-headless>=4.1.1  # 专为服务器环境设计的 OpenCV，不带 GUI，更稳
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
matplotlib>=3.2.2
flask>=2.0.0

# --- 工具依赖 ---
psutil  # 系统监控
thop>=0.1.1  # 计算 FLOPs